pyproject.toml
[project]
name = "cross-agentic-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.6.1",
    "fastapi>=0.128.5",
    "mistral>=21.0.0",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
